[Ref] Move lucene dao to parent dao package
